# fix(es/minifier): Avoid paren when compress tenary