--- a/crates/swc_ecma_minifier/tests/fixture/issues/react-instancesearch/004/output.js
+++ b/crates/swc_ecma_minifier/tests/fixture/issues/react-instancesearch/004/output.js
@@ -21,7 +21,7 @@ var _obj, isMultiIndexContext = function(widget) {
     return widget.props.indexId === indexId;
 }, sortIndexWidgetsFirst = function(firstWidget, secondWidget) {
     var isFirstWidgetIndex = isIndexWidget(firstWidget), isSecondWidgetIndex = isIndexWidget(secondWidget);
-    return isFirstWidgetIndex && !isSecondWidgetIndex ? -1 : +(!isFirstWidgetIndex && !!isSecondWidgetIndex);
+    return isFirstWidgetIndex && !isSecondWidgetIndex ? -1 : !isFirstWidgetIndex && isSecondWidgetIndex ? 1 : 0;
 };
 /**
  * Creates a new instance of the InstantSearchManager which controls the widgets and