… crates/swc_ecma_minifier/tests/benches-full/victory.js
… crates/swc_ecma_minifier/tests/fixture/next/syncfusion/933-e9f9a6bf671b96fc/output.js
… crates/swc_ecma_minifier/tests/full/issue-5912-bigdecimal/output.js
… crates/swc_ecma_minifier/tests/projects/output/jquery.mobile-1.4.2.js