@kdy1 2025-01-22
# Main:

```
    Finished `release` profile [optimized + debuginfo] target(s) in 0.15s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/antd.js`
File: ./benches/full/antd.js
  456868
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/d3.js`
File: ./benches/full/d3.js
   86866
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/echarts.js`
File: ./benches/full/echarts.js
  319497
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/jquery.js`
File: ./benches/full/jquery.js
   30875
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/lodash.js`
File: ./benches/full/lodash.js
   25030
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/moment.js`
File: ./benches/full/moment.js
   18572
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/react.js`
File: ./benches/full/react.js
    8185
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/terser.js`
File: ./benches/full/terser.js
  116687
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/three.js`
File: ./benches/full/three.js
  157550
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/typescript.js`
File: ./benches/full/typescript.js
  806012
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/victory.js`
File: ./benches/full/victory.js
  157110
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/master-swc/target/release/examples/minifier ./benches/full/vue.js`
File: ./benches/full/vue.js
   42412
~/projects/master-swc/crates/swc_ecma_minifier main 8s
```


This PR:

```
    Finished `release` profile [optimized + debuginfo] target(s) in 0.17s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/antd.js`
File: ./benches/full/antd.js
  456868
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/d3.js`
File: ./benches/full/d3.js
   86865
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/echarts.js`
File: ./benches/full/echarts.js
  319492
    Finished `release` profile [optimized + debuginfo] target(s) in 0.15s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/jquery.js`
File: ./benches/full/jquery.js
   30872
    Finished `release` profile [optimized + debuginfo] target(s) in 0.15s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/lodash.js`
File: ./benches/full/lodash.js
   25025
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/moment.js`
File: ./benches/full/moment.js
   18572
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/react.js`
File: ./benches/full/react.js
    8185
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/terser.js`
File: ./benches/full/terser.js
  116687
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/three.js`
File: ./benches/full/three.js
  157549
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/typescript.js`
File: ./benches/full/typescript.js
  806002
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/victory.js`
File: ./benches/full/victory.js
  157110
    Finished `release` profile [optimized + debuginfo] target(s) in 0.14s
     Running `/Users/kdy1/projects/s/pr/target/release/examples/minifier ./benches/full/vue.js`
File: ./benches/full/vue.js
   42412
~/projects/s/pr/crates/swc_ecma_minifier Austaras/main 8s
```

So this is a clear win for the bundle size. I'll try to understand logic after taking some sleep...